requirements.txt
# Core dependencies for AI Social Network Analysis
praw>=7.7.0
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.31.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Dashboard
streamlit>=1.28.0

# Optional but recommended
python-dotenv>=1.0.0

# Development dependencies (testing, linting, type checking)
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
mypy>=1.4.0
black>=23.0.0
isort>=5.12.0
